+++ mono/src/main/java/woowa/team4/bff/event/menu/MenuCreateEvent.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.event.menu;
+
+public record MenuCreateEvent(Long menuId, Long restaurantId, String menuName) {
+}
