+++ mono/src/main/java/woowa/team4/bff/event/menu/MenuUpdateEvent.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.event.menu;
+
+public record MenuUpdateEvent(Long menuId, Long restaurantId, String menuName) {
+}
