+++ mono/src/main/java/woowa/team4/bff/event/restaurant/RestaurantCreateEvent.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.event.restaurant;
+
+public record RestaurantCreateEvent(Long restaurantId, String restaurantName) {
+}
