+++ mono/src/main/java/woowa/team4/bff/event/restaurant/RestaurantUpdateEvent.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.event.restaurant;
+
+public record RestaurantUpdateEvent(Long restaurantId, String restaurantName) {
+}
