+++ mono/src/main/java/woowa/team4/bff/search/command/CreateRestaurantSearchCommand.java
@@ -1,4 +1,4 @@
-package woowa.team4.bff.search.service.command;
+package woowa.team4.bff.search.command;
 
 public record CreateRestaurantSearchCommand(Long restaurantId, String restaurantName) {
     public static CreateRestaurantSearchCommand of(Long restaurantId, String restaurantName) {
