+++ mono/src/main/java/woowa/team4/bff/search/command/SearchRestaurantCommand.java
@@ -1,4 +1,4 @@
-package woowa.team4.bff.search.service.command;
+package woowa.team4.bff.search.command;
 
 public record SearchRestaurantCommand(String keyword, String deliveryLocation) {
     public static SearchRestaurantCommand of(String keyword, String deliveryLocation) {
