+++ mono/src/main/java/woowa/team4/bff/search/command/UpdateRestaurantSearchCommand.java
@@ -1,4 +1,4 @@
-package woowa.team4.bff.search.service.command;
+package woowa.team4.bff.search.command;
 
 public record UpdateRestaurantSearchCommand(Long restaurantId, String restaurantName) {
 }
