+++ mono/src/main/java/woowa/team4/bff/search/service/SearchIndexManageService.java
@@ -4,10 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
+import woowa.team4.bff.event.menu.MenuCreateEvent;
+import woowa.team4.bff.event.menu.MenuUpdateEvent;
+import woowa.team4.bff.event.restaurant.RestaurantCreateEvent;
+import woowa.team4.bff.event.restaurant.RestaurantUpdateEvent;
+import woowa.team4.bff.search.domain.MenuSearch;
 import woowa.team4.bff.search.domain.RestaurantSearch;
 import woowa.team4.bff.search.repository.SearchIndexManageRepository;
-import woowa.team4.bff.search.service.command.CreateRestaurantSearchCommand;
-import woowa.team4.bff.search.service.command.UpdateRestaurantSearchCommand;
 
 @Service
 @RequiredArgsConstructor
@@ -15,20 +18,32 @@ public class SearchIndexManageService {
     private final SearchIndexManageRepository searchIndexManageRepository;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public Long addRestaurant(CreateRestaurantSearchCommand command) {
+    public Long addRestaurant(RestaurantCreateEvent event) {
         RestaurantSearch restaurantSearch = RestaurantSearch.builder()
-                .restaurantId(command.restaurantId())
-                .restaurantName(command.restaurantName())
+                .restaurantId(event.restaurantId())
+                .restaurantName(event.restaurantName())
                 .build();
         return searchIndexManageRepository.save(restaurantSearch);
     }
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public Long updateRestaurant(UpdateRestaurantSearchCommand command) {
-        RestaurantSearch restaurantSearch = searchIndexManageRepository.findByRestaurantId(command.restaurantId());
-        restaurantSearch.update(command.restaurantName());
+    public Long updateRestaurant(RestaurantUpdateEvent event) {
+        RestaurantSearch restaurantSearch = searchIndexManageRepository.findByRestaurantId(event.restaurantId());
+        restaurantSearch.update(event.restaurantName());
         return searchIndexManageRepository.save(restaurantSearch);
     }
 
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public Long addMenu(MenuCreateEvent event) {
+        MenuSearch menuSearch = MenuSearch.builder()
+                .build();
+        return searchIndexManageRepository.save(menuSearch);
+    }
 
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public Long updateMenu(MenuUpdateEvent event) {
+        MenuSearch menuSearch = searchIndexManageRepository.findByMenuId(event.menuId());
+        menuSearch.update(event.menuName());
+        return searchIndexManageRepository.save(menuSearch);
+    }
 }
